Revert "finalizing networking"
This reverts commit 99b84b5e96399bcc4eb8fc524353530911dd7cff.
