+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Wed Oct 23 18:24:54 EDT 2019
+gradle.version=5.4.1
