+++ app/build.gradle
@@ -44,6 +44,5 @@ dependencies {
     implementation 'net.sourceforge.streamsupport:streamsupport:1.7.0'
     implementation "com.squareup.moshi:moshi-adapters:1.8.0"
     implementation "com.squareup.moshi:moshi-kotlin:1.8.0"
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.2'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.2'
+
 }
