+++ app/src/main/java/com/example/ithaca_transit_android_v2/Models/ActionType.kt
@@ -1,4 +1,4 @@
-package com.example.ithaca_transit_android_v2.models
+package com.example.ithaca_transit_android_v2.Models
 
 /*
 Represents the types of action a user can take while on their Trip.
