+++ app/src/main/java/com/example/ithaca_transit_android_v2/Models/Location.kt
@@ -1,5 +1,6 @@
 package com.example.ithaca_transit_android_v2.models
 
+import com.example.ithaca_transit_android_v2.Models.LocationType
 import com.squareup.moshi.JsonClass
 
 //Location. - Represents either a bus stop or a google place.
