+++ app/src/main/java/com/example/ithaca_transit_android_v2/Models/LocationType.kt
@@ -1,0 +1,6 @@
+package com.example.ithaca_transit_android_v2.Models
+
+enum class LocationType {
+    busStop,
+    googlePlace
+}
