+++ app/src/main/java/com/example/ithaca_transit_android_v2/Models/Route.kt
@@ -1,4 +1,6 @@
-package com.example.ithaca_transit_android_v2.models
+package com.example.ithaca_transit_android_v2.Models
+import com.example.ithaca_transit_android_v2.models.Coordinate
+import com.example.ithaca_transit_android_v2.models.Location
 import java.util.*
 
 data class Route(
@@ -9,4 +11,4 @@ data class Route(
     val endLocation: Location,
     val busStops: List<Location>,
     val busNumber: Int
-)
+) {}
