+++ app/src/main/java/com/example/ithaca_transit_android_v2/Models/RouteType.kt
@@ -1,4 +1,4 @@
-package com.example.ithaca_transit_android_v2.models
+package com.example.ithaca_transit_android_v2.Models
 
 /*
 Represents the type of route that is displayed on the map
