+++ app/src/main/java/com/example/ithaca_transit_android_v2/Models/Trip.kt
@@ -1,5 +1,6 @@
-package com.example.ithaca_transit_android_v2.models
+package com.example.ithaca_transit_android_v2.Models
 
+import com.example.ithaca_transit_android_v2.models.Location
 import java.util.Date
 
 // [Trip] is a collection of [Route] objects and other essential information
