--- app/src/main/java/com/example/ithaca_transit_android_v2/models/LocationType.kt
@@ -1,6 +1,0 @@
-package com.example.ithaca_transit_android_v2.models
-
-enum class LocationType {
-    BUS_STOP,
-    GOOGLE_PLACE
-}
